requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.7.0
Pillow>=9.0.0
transformers>=4.31.0
torch>=2.0.0
numpy>=1.23.0
requests>=2.31.0